+++ tests/test_invalid_actions.py
@@ -68,7 +68,8 @@ def test_eval_env():
     env = InvalidActionEnvDiscrete(dim=20, n_invalid_actions=10)
     eval_env = InvalidActionEnvDiscrete(dim=20, n_invalid_actions=10)
     model = MaskablePPO("MlpPolicy", env, clip_range_vf=0.2, n_steps=32, seed=8)
-    model.learn(32, eval_env=eval_env, eval_freq=16)
+    with pytest.warns(DeprecationWarning):  # `eval_env` is deprecated
+        model.learn(32, eval_env=eval_env, eval_freq=16)
     model.learn(32, reset_num_timesteps=False)
 
 
@@ -189,7 +190,7 @@ def test_callback(tmp_path):
     model = MaskablePPO("MlpPolicy", env, n_steps=64, gamma=0.4, seed=32, verbose=1)
     model.learn(100, callback=MaskableEvalCallback(eval_env, eval_freq=100, warn=False, log_path=tmp_path))
 
-    model.learn(100, callback=MaskableEvalCallback(Monitor(eval_env), eval_freq=100, warn=False))
+    model.learn(100, callback=MaskableEvalCallback(Monitor(eval_env), eval_freq=100, warn=False), progress_bar=True)
 
 
 def test_child_callback():